Simplify upgrades system
Note that I have removed the speed upgrade from the UpgradeType enum. In
my opinion, because the speed upgrade is non-droppable, it should not be
considered an UpgradeType. Perhaps speed upgrades could be directly
applied to the player from the future code that handles the upgrades
store.

Also, my code is formatted slightly differently from the rest of the
code base, which is written in a style more reminiscent of C++ than
Java. I can fix this if necessary.